Refactoring of StagingSession without mode layer
