+++ MediaCenter/MediaCenter/Sessions/SessionBase.cs
@@ -1,16 +1,16 @@
-﻿using MediaCenter.Helpers;
-using MediaCenter.MVVM;
-using MediaCenter.Repository;
+﻿//using MediaCenter.Helpers;
+//using MediaCenter.MVVM;
+//using MediaCenter.Repository;
 
-namespace MediaCenter.Sessions
-{
-    public abstract class SessionBase : PropertyChangedNotifier
-    {
-        public readonly IRepository Repository;
+//namespace MediaCenter.Sessions
+//{
+//    public abstract class SessionBase : PropertyChangedNotifier
+//    {
+//        public readonly IRepository Repository;
         
-        protected SessionBase(IRepository repository)
-        {
-            Repository = repository;
-        }
-    }
-}
+//        protected SessionBase(IRepository repository)
+//        {
+//            Repository = repository;
+//        }
+//    }
+//}
